DeviceInformation.CreateWatcher(String) not available in JS
The DeviceInformation.CreateWatcher(String) method is not available in JS. The single-parameter overload uses the CreateWatcher(DeviceClass) overload, and the string will be coerced to an integer. This will typically coerce to 0 which corresponds to DeviceClass.All.

To obtain AQS functionality, use the CreateWatcher(String, IIterable<String>) overload and pass an empty iterable.