Add a reference to ContentDialog in Popup.md
MessageDialog is a Windows 8 control, whereas the ContentDialog control has effectively replaced it for UWP apps. This is made clear in a note in the MessageDialog page, but that adds an extra hop to find the control the person searching is actually looking for, and makes for an extra context switch. Instead of needing to follow multiple redirections, by adding a reference to ContentDialog to the Popup page, and disambiguating when each control should be used, this change makes finding information about dialogs easier.