Change in MapControl description
The MapControl class doesn't represent a MapControl--it represents a map.  I was tempted to put "3D", but it has 2D views also.  I was tempted to put "interactive", but other controls are interactive without stating it.  So I settled on the fact that you can see it with imagery or just symbolically.