Merged PR 3085: 13400476 Docs.ms.com June2017 Customer Feedback | LanIdentifier Class | req for
The customer asked for a code example for this type. There actually is a link to a code example there already but a) it was clearly not prominent enough on the page, so I've moved it to the head of the page, b) the examples section was empty so I've put the link there, too, and c) the link was to a topic outside the UWP docset proper (and it was a HTML one) so I've updated the url.