Update inkstroke_id.md
I counted on the documentation about the Id not changing through serialization/deserialization and found that to not be true. It also appears that it restarts the counter (at 1) upon load. I believe that the ID being unique only applies to a particular InkContainer instance. It would be nice to have some clarification about what is unique about the ID...