Update packagemanager_findpackageforuser_526853699.md
Update "This user security ID of the user for whom the specified package is being retrieved. If this parameter is null, the specified package is retrieved for the current user."
with 
"The user security identifier (SID). If this parameter is string.Empty, the packages are retrieved for the current user."
This has been approved by a SME. Thanks!