Merged PR 4777: Resolve apparent contradiction in UEFI/TPM requirement for SystemIdentification.GetSystemIdForPublisher
Resolve apparent contradiction in UEFI/TPM requirement for SystemIdentification.GetSystemIdForPublisher

The paragraph originally said that the method requires UEFI or TPM. But then it later talks about what it does if you don't have one, which contradicts the previous statement that it requires a UEFI or TPM.

Remove the requirement. The requirement was in place for build 14393 (RS1 = Anniversary Update = 1607) but was lifted in build 16299 (RS3 = Fall Creators Update = 1709).