Merged PR 8152: Clarify ClaimedReceiptPrinter.LinesToPaperCut
The point is that this is a static value that tells you the recommended number of additional blank lines to place at the end of your receipt so that it clears the cutting blade. It is not a dynamic value that tells you how many lines you still need to advance.