Merged PR 8808: Correct C++/WinRT documentation for DateTime/TimeSpan.
C++/WinRT projects Windows.Foundation.DateTime and TimeSpan as specializations of std::chrono::time_point and std::chrono::duration, respectively. As a result, they don't have the same field names as the other projections, and users should instead use the std::chrono API to interact with them.

Also, see this question on stackoverflow that alerted me to the documentation discrepancy: https://stackoverflow.com/questions/50658795/how-to-get-the-current-time-in-utc