Clarify the PausedSystemPolicy values (!phone)
The docs used to say that the PausedSystemPolicy value could only be returned on Windows Phone. Actually, the behavior was converged in TH1. The documentation should not be explicit about exactly when things get paused because (a) the NDX team wants to be able to change the exact times and (b) the system team wants to be able to turn background tasks on and off at will and definitely want to make more changes in the future.