Space was causing a reference link not to get rendered
In the forth paragraph of Remarks there was a portion of the content that were not rendering the link to a reference. appearing this way
" Like the@Windows.Foundation.Diagnostics.LoggingSession?text=LoggingSession class, LoggingChannel instances are added to a FileLoggingSession instance, and the FileLoggingSession instance has methods to remove channels and dispose." where the @Windows.Foundation..... should have been a link but this was part of the previous word "the" and this was causing the @ reference tag not to be properly parsed.