Merged PR 11216: Clarifying ImageOpened and ImageFailed events
A partner team hit an issue where they expected ImageOpened and ImageFailed to fire to do some logic. However, ImageOpened and ImageFailed didn't fire when expected. The culprit was that the image wasn't in the live tree. We expect that other customers might hit this issue, so adding a small note to the two events that describes the "by design" behavior.