Removed migration info from Windows 8 to 8.1.
The [BitmapImage] class described here is for the UWP platform which is exclusive to Windows 10. However, there is an entire paragraph dedicated to porting an app using an Image object from Windows 8 to Windows 8.1.

More than 3 years into Windows 10's life, I guess it is reasonable to assume that no one is porting apps from Windows 8 to Windows 8.1 or to Windows 10 any longer. Hence, any such reference to Windows 8 (8.1) should be removed from this UWP documentation as it is merely a big wall of unnecessary information.