Merged PR 12332: Add a cross-link back to the main page for XamlDirect
Add a cross-link back to the main page for XamlDirect.

The linked-to page has many links to the changed page, but no links exist in the other direction. This makes it hard to find the parent, which actually describes XamlDirect.

Perhaps it might make sense to remove most of the links from https://docs.microsoft.com/en-us/uwp/api/windows.ui.xaml.core.direct to the changed page, since I clicked most of them expecting more information about the product, rather than an API documentation page. If we're using "XamlDirect" to mean "the set of APIs that are called XamlDirect," then the user is already on the correct page. If we're using `XamlDirect` to mean "the namespace for XamlDirect", then it is appropriate to link.