Merged PR 13959: Updated navigationviewitem.md to edit remark that was only true prior to v1809(17763)
Updated navigationviewitem.md to share that prior to Windows 10, version 1809 (10; Build 17763) you can only place NavigationViewItems in NavigationView.MenuItems.
In 1809, changes were made to allow NavigationViewItems elsewhere.
Please note, "Windows 10, version 1809 (10; Build 17763)" is how the Target version is specified in Visual Studio project files. If this should be changed, please let me know, but I chose it because that's how we present it in Visual Studio and both Windows 10 1809 and Target SDK for 17763 are both required if you place NavigationViewItems anywhere other than the MenuItems property.