Include info on which XML this accepts and when
We fixed the method in the Windows 10 May 2019 Updates (version 1903), build 18362.267 (the July 27 2019 update).
The XML was incorrectly named.

Note that this change requires an editor to make sure that (a) the link is fixed to match the actual markdown requirements and (b) that the remarks are properly formatted. There's one exact intermediate release at which point this method started to work; before then it didn't.