Bias-Free Communications
Microsoft supports a diverse and inclusionary environment.  Within this document, there is a reference to the word 'slave.' Microsoft's [Style Guide for Bias-Free Communications](https://github.com/MicrosoftDocs/microsoft-style-guide/blob/master/styleguide/bias-free-communication.md) recognizes this as an exclusionary word.  The change I made edits the word 'slave' to 'subordinate' to be in line with the Style Guide.