Merged PR 17215: Add C++/WinRT example for UIElement.AddHandler
As in PR title, adding an example for C++/WinRT.
I see that the snippets for other languages come from some project files which I assume compile, but I didn't see one for cpp-winrt so I just added the snippet in-place.
Let me know if there's a better way to do it.

PointerPressed is handled synchronously, so using a raw `this` pointer should be fine.

Removing the C++/CX sample, as it is discouraged in favor of cppwinrt.