Document the mystery empty contact that GetMeContact returns
Also, call it "the contact that represents the current user"
rather than "the contact for the current user", because "for the current
user" is a Phrase of Power that has MUA connotations.

(The guidance for checking the Id seems correct, but there
is a remark on Contact.Id that says that you can indeed add
a contact with an empty Id. You won't be able to find it later,
though, and the implementation suggests that this means that
the system will autogenerate a nonempty ID.)