Merged PR 17357: Fix widespread incorrect descriptions of "Auto" aka NaN in XAML documentation
In a handful of places, XAML allows a floating point value to be NaN, which is used as a sentinel value to represent some special (usually default) behavior.

Existing text was very C#-centric. Made it more language-neutral and put language-specific details into language-specific NOTEs. (I couldn't figure out how to conditionalize text based on the user's preferred programming language, so I just put both of them in there, with a callout for which language it applies to.)

Existing text erroneously said that C++ cannot generate NaN. In fact, it can, in many different ways! Describe two common ways to get NaN in C++.

Removed confusing text that talks about how the NaN value prints out, which is not really relevant, and is also a quirk of the Microsoft compiler, which is not standard-conforming in this respect. The standard C++ behavior is for NaN to print as "nan" or "nan(...)" where "..." is implementation-defined text.

Existing text about NaN said that in XAML markup, you can write "Auto" as an alias for NaN. This is true for FrameworkElement, but nowhere else. Don't call it "Auto" in places where it isn't supported, and don't say that XAML lets you write "Auto" when it doesn't.

Existing text about NaN said that "Auto" means to "size to available size in layout". This is true only for FrameworkElement, so remove that text from everywhere else. Replaced with the *actual* behavior of NaN, based on existing text or personal research. Please validate that my descriptions of the actual behavior are correct.

The ones I'm least confident about are ItemsWrapGrid and WrapGrid. I found the source for WrapGrid, which talks about "realized items" in some funky way. I couldn't find where ItemsWrapGrid uses the ItemHeight/Width properties. I based the descriptions on experimentation.

I also couldn't find where the RasterizePixelHeight/Width are used in SvgImageSource, but the existing description seemed precise enough that I'm inclined to believe it.