Merged PR 17644: ListView and GridView API changes for WinUI 2.6
There were some changes made to the ListView and GridView APIs in the WUX namespace for the WinUI 2.6 release, I've updated the API docs to reflect those.

This shouldn't be merged until the WinUI 2.6 release. I'm waiting on a link to the release notes then my PR will be complete.