Merged PR 18422: Add remark to Radio.SetStateAsync that it may throw ERROR_BUSY on completion
It was discovered that due to the lower-level implementation, callers may get an ERROR_BUSY exception when attempting to get the result of the operation from the `IAsyncOperation<RadioAccessStatus>` returned by `Radio.SetStateAsync()`.

This case in uncommon and occurs only when another process is already changing the radio state. We haven't seen evidence of this implementation detail impacting partners or end users, but documenting this case allows developers to safeguard their applications from this issue if it is a concern.

Added a remark noting that the ERROR_BUSY exception may be thrown, that it should be caught and the operation retried.