Update MoveAsync with exception details
Update StorageFile.MoveAsync method definitions to provide more details about the exception thrown if the destinationFolder param is not a physical folder on disk.
This request originated from a community GitHub issue:
https://github.com/MicrosoftDocs/winrt-api/issues/1072

Let me know if this is the correct formatting for System.Exception and the error message text. Should they be inline code style instead?